Add StartDialog function
This commit also changed the cache global variable's name from DLG_SELF
to DIALOG. DLG_SELF is now used to denote the NPC speaker in the dialog.
It is equivalent to OBJECT_SELF in most cases, but in a conversation
with an object other than a creature or placeable, OBJECT_SELF will be
the PC, so DLG_SELF can be used to get the object the conversation is
supposed to be with.