Properly set DLG_SELF on dialog init
Normally, OBJECT_SELF can be used to get the object the PC is conversing
with. However, if the conversation was triggered with an object other
than a creature or placeable, OBJECT_SELF will actually be the PC
himself. When a conversation is initiated, DLG_SELF is set to the object
the PC is conversing with, so it can reliably be used to get that
object.