Add hex, RGB, and HSV color utilities
Added functions to convert RGB, HSV, and hex color codes into NWN color
tags. Hex codes will be the preferred method when coloring strings, as
they can be easily looked up and copied and pasted. The color constants
have been changed to equivalent hex colors to reflect this.

Note: the HSV functions have not been thoroughly tested, and there may
be some bugs. I've noted that the values they yield when converting to
other formats are slightly off due to the nature of floating point
arithmetic. The differences in color should not be enough to really
notice.