Add dialog token caching
You can now cache the values of dialog tokens to prevent re-running
their eval scripts when you know the value will be the same. You can do
this in one of two ways:

- Automatically: in the token eval script, set a local int on the PC
  with the varname `*TokenCache` and a value of TRUE. The returned value
  will then be captured and cached.
- Manually: in the dialog script, use the CacheDialogToken() function to
  set a value for the token. The token eval script will not run for this
  token until the cache is cleared with UnCacheDialogToken().

Cached token value are specific to the token's capitalization. For
example, caching a value for <his/her> will not affect the token
<His/Her>. This allows tokens to cache a properly capitalized value.

Usually, a cached token value will last for the lifetime of the dialog.
However, if you manually set the token value in the dialog's init
script, the value will remain for all future instances of the dialog.
That's probably not something you want if the value of the token should
depend on the PC or game state at the time the dialog is intantiated.