Improve programatic plugin creation
This update makes it easier to create plugins programatically. Now, a
list of libraries is loaded first. The libraries can use their
OnLibraryLoad() routines to create plugin objects and set their names,
descriptions, and event scripts. Next, a list of plugins is loaded. If
the plugin was not defined by a library, it can be created from an item
blueprint.