Add item-specific (un)acquire and (un)equip events
Items can now specify their own OnAcquireItem, OnUnacquireItem,
OnPlayerEquipItem, OnPlayerUnequipItem, and OnActivateItem events. This
works by:

1. Running the event on the module
2. Running the event locally on the item
3. Running the tag-based script if neither of these events issues a
   DENIED state.

This means you can add a string variable named OnActivateItem to an item
that causes that script to be launched when that particular item is
activated. This can be used in place of tag-based scripts.

Note that these locally run scripts treat the item rather than the
module as OBJECT_SELF.