Add option to run a pre-load script before framework initialization (#21)
Generally this is not an issue in EE, however, in 1.69, initializatin the core-framework can cause TMI very quickly due to the modular nature of scripts and plugins that the framework allows. This option was added in order to allow the build to run a specific script before the framework initialized. Specifically, this was designed to allow NWNX2 users to raise thier TMI limit through nwnx_tmi before running into TMI during framework initilializtion, but it can be used for any functions the builder desires.