refactor: remove nwnx-specific code
All nwnx-specific code has been removed from the core files except for
hook_nwnx.nss. To build with nwnx, users will need to have downloaded
the nss files from the nwnxee github releases page and place them in a
directory called nwnxee in the same directory as the framework. If the
user does not want nwnx, they can comment out the lines that do so in
nasher.cfg.