feat: automatically hook all objects on init
This adds a setting to automatically hook any object in the module when
the Core Framework is initialized. The `AUTO_HOOK_OBJECT_EVENTS` can be
bitmasked to include all object types the builder wants to automatically
hook. Similarly, `AUTO_HOOK_OBJECT_HEARTBEAT_EVENT` can include all
object types the builder wants to hook the heartbeat event for.

To demonstrate this functionality, all references to "hook_nwn" have
been removed from the demo module except for the OnModuleLoad and
OnAreaEnter events.