feat!: automatically load libraries by pattern
The `INSTALLED_LIBRARIES` constant now supports string patterns as well
as literal library names. This allows the pattern `**_l_plugin` to be
included in the list of libraries wherever the user wants instead of
being a simple toggle. This change removes the `AUTO_LOAD_PLUGINS`
setting, so `INSTALLED_LIBRARIES` must be updated to include the plugin
pattern. This is a BREAKING CHANGE.