feat: add default conversation handler to bw_defaultevents
The default conversation handler nw_g0_conversat is fired on a
placeable, object, or creature that does not have a conversation script
set. If the Core Framework handles these events, that script does not
fire. This change adds nw_g0_conversat as a default event script, which
means it will not fire if one has already been set for the event. This
allows placeables and doors to trigger a conversation when used. It
should not fire for creatures, since they have the nw_c2_default4 script
as well.