refactor!: use glob matching to load libraries
The Core now uses the new sm-utils version of LoadLibrariesByPattern()
which uses glob syntax instead of TestStringAgainstPattern(). It also
takes a CSV list of patterns and loads the libraries in the order they
match the patterns, allowing finer control over the load order.

However, this is a BREAKING CHANGE. Sorry if you built your module on
that during the whole day the previous version was up.