feat: add plugin for util_i_targeting events (#40)
This adds a basic plugin (much like the chat plugin) that has two
targeting-specific hooks: one for OnModuleLoad to create the targeting
tables, and one for OnPlayerTarget to check if the targeting player has a global
targeting hook assigned. If a targeting hook is assigned, further scripts are
not run.