add default scripts into HookObjectEvents()
To remedy a situation where objects with default scripts, specifically
objects with transitions, quit working after being hooked by the
framework, I added the default scripts into HookObjectEvent().

The reason all of the default scripts are included here, including
`nw_g0_conversat` and `nw_g0_transition`, even though they are included
as variables on the `bw_defaultevents` plugin item, is because this
behavior is mandatory to ensure the transitions and conversations work
correctly, while the inclusion of the `bw_defaultevents` item is not
mandatory as it's part of the framework demo and not the core.  The idea
here is that system should still run correctly if no files from folders
other than `core` are included.

I also added the `nw_g0_transition:default` script to the appropriate
events on `bw_defaultevents.uti`.

Finally, I wasn't creative enough to find a way to add default
`OnTrapTriggered` behavior (2da lookup) to the `bw_defaultevents` item,
so that functionality only exists in `HookObjectEvent()`.